Architecture refactoring
