--- manager/filesystem.go
@@ -1,82 +1,0 @@
-package manager
-
-import (
-	"bufio"
-	"fmt"
-	"os"
-	"path/filepath"
-	"strconv"
-	"strings"
-
-	"github.com/gh0st17/timetable-go/errtype"
-)
-
-// Функция для записи строки в файл
-func writeString(filePath string, data *string) error {
-	var (
-		file *os.File
-		err  error
-	)
-
-	if file, err = os.Create(filePath); err != nil {
-		return errtype.ErrRuntime(fmt.Errorf("ошибка создания файла %s: %s", filePath, err))
-	}
-	defer file.Close()
-
-	if _, err = file.WriteString(*data); err != nil {
-		return errtype.ErrRuntime(fmt.Errorf("ошибка записи в файл %s: %s", filePath, err))
-	}
-
-	return nil
-}
-
-// Печать строк с нумерацией
-func printLines(lines []string, p *Params, printOnly bool) {
-	if printOnly {
-		for _, line := range lines {
-			fmt.Println(line)
-		}
-	} else {
-		fmt.Printf("Группы %d курса института №%d\n\n", p.Course, p.Dep)
-		for i, line := range lines {
-			fmt.Printf("%d. %s\n", i+1, line)
-		}
-	}
-}
-
-// Обработка пользовательского ввода
-func getUserSelection(lines []string) uint64 {
-	var (
-		err    error
-		input  string
-		result uint64
-	)
-
-	for {
-		fmt.Printf("\nВыберите номер группы в списке: ")
-		input, err = bufio.NewReader(os.Stdin).ReadString('\n')
-		if err != nil {
-			fmt.Println("Ошибка ввода:", err)
-			continue
-		}
-
-		input = strings.TrimSpace(input)
-		result, err = strconv.ParseUint(input, 10, 64)
-		if err != nil || result < 1 || result > uint64(len(lines)) {
-			fmt.Println("Неверный ввод. Попробуйте снова.")
-			continue
-		}
-
-		fmt.Println()
-		return result - 1
-	}
-}
-
-// Функция для получения абсолютного пути запускаемой программы
-func getWd() (string, error) {
-	if executable, err := os.Getwd(); err != nil {
-		return "", err
-	} else {
-		return filepath.Abs(executable)
-	}
-}
