+++ manager/internal/database/err.go
